Keep computed font weight resolution with its property
Inline the one-use helper into fontWeight.js and retain the TODO documenting unresolved special cases in the generic computed-value path.
