+++ lib/jsdom/living/css/CSSStyleDeclaration-impl.js
@@ -480,6 +480,8 @@ class CSSStyleDeclarationImpl {
       }
     }
 
+    // TODO: Resolve special cases other than color.
+
     return value;
   }
