+++ lib/jsdom/living/css/properties/fontWeight.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const parsers = require("../helpers/css-values");
-const { resolveFontWeight } = require("../helpers/font-weights");
 
 const property = "font-weight";
 const shorthand = "font";
@@ -29,13 +28,45 @@ const descriptor = {
 };
 
 function resolveComputedValue(value, { getInheritedValue, resolveDefault }) {
-  return resolveFontWeight(value, {
-    getInheritedWeight: () => getInheritedValue(property, {
-      inherit: true,
-      initial: "400"
-    }),
-    resolveDefault
+  if (value === "normal") {
+    return "400";
+  }
+  if (value === "bold") {
+    return "700";
+  }
+  if (value !== "bolder" && value !== "lighter") {
+    return resolveDefault(value);
+  }
+
+  const inheritedValue = getInheritedValue(property, {
+    inherit: true,
+    initial: "400"
   });
+  const inheritedWeight = Number(inheritedValue);
+
+  if (value === "bolder") {
+    if (inheritedWeight < 350) {
+      return "400";
+    }
+    if (inheritedWeight < 550) {
+      return "700";
+    }
+    if (inheritedWeight < 900) {
+      return "900";
+    }
+    return inheritedValue;
+  }
+
+  if (inheritedWeight < 100) {
+    return inheritedValue;
+  }
+  if (inheritedWeight < 550) {
+    return "100";
+  }
+  if (inheritedWeight < 750) {
+    return "400";
+  }
+  return "700";
 }
 
 /**
